One worker for all, many workers for each, perform_bulk
